Fixed: ignore step checker, before updating profile version.
When the last applied upgrade step had a checker, the profile version
was not updated.  Now we no longer look at the checker of the last
applied step when deciding whether to set the profile version.  The
checker, if any is set, normally returns True before running the step
(it can be applied), and False afterwards (it was already applied).