Be forgiving with profile ids with or without 'profile-' at the start.
All functions that accept a profile id argument and only work when the
id does *not* have this string at the start, will now simply strip it
off if it is there.  For example, 'getLastVersionForProfile' will give
the same answer whether you ask it for the version of profile id
'foo:default' or 'profile-foo:default'.

The various registries are storing 'foo:default' as profile id, but
various functions return 'profile-foo:default' when reporting about
this profile.

Now you don't have to double or triple check yourself.