Fix 126: Handle legacy versions in upgrades also with setuptools 66 or higher.
With setuptools 66+, when parsing fails, we retry with `0+<legacy version>`
With earlier setuptools we still get a LegacyVersion if the version cannot be stricly parsed, so this part does not change.

In all cases, if we get an empty version or the special version `unknown`, we treat it as zero.